Complete Palindrome Number (leetcode #9)
If the input is negative, return false.
First solution was to turn x into a string, reverse the string,
and compare the original string to the reversed copy.

To save string translation runtime, use math instead to build
the reverse number. Reverse set to 0. Copy original input. While
the copy is greater than 0, reverse is equal to 10 * itself +
the rightmost digit of copy (copy % 10). Divide copy by 10 and
continue building. Compare x to the reverse.