Roman to int (#2)
* Solve Roman to integer LeetCode #12

Iterate through the string building the result and return the
integer value at the end.

* Refactor roman to int for simplicity

This change adds a helper function for converting a character
from roman numeral to integer and more elegantly handles the
result generation. Instead of using the switch clause with if
cases to handle the individual cases where a smaller numeral is
ahead of a larger one, iterate through the string directly comparing
each character with the one that follows it (if one exists) and
add or subtract accordingly.

---------

Co-authored-by: dgralla <dgralla@github.com>