Solve two sum (problem #1)
Use a hashmap to store the difference between the target
and the current value, mapped to its index. If the current
value is already in the map, then a previous number sums to
the target with the current value. Return the int array of
the index from the map and the current index.

Also included a simple intArrayToString for visualizing the
output.