Solve longest substring without repeating characters
This solution uses the sliding window approach in order to run
in linear time O(n). Initialize a left pointer and a max length
variable to 0 and create an empty set of characters. Then add
characters to the set by moving right through the string, unless
that character already exists in the set. If the character exists
in the set, empty the set, slide the left pointer to equal the right,
and start the next substring from there. Return the max window size
where none of the characters repeated.