Solve longest common prefix
This commit includes two solutions: the commented out solution
iterated through the list tracking the longest common prefix thus
far (using the commonPrefix helper), and breaking whenever there
is no longer a common prefix. If the end of the list is reached,
return the result.

The uncommented solution sorts the array of strings, which is done
alphabetically by default. Then find and return the common prefix
between the first and last strings in the sorted array, noting that
all the strings in between must share that prefix as well.